biology script

+++ evaluation/benchmarks/bio/setup.sh
@@ -8,9 +8,12 @@ mkdir -p $PW
 mkdir -p ${PASH_TOP}/evaluation/bio/bio4/output
 cat ./input.txt |while read s_line;
 	do
+    
     sample=$(echo $s_line |cut -d " " -f 2);
-    pop=$(echo $s_line |cut -f 1 -d " ");
-    link=$(echo $s_line |cut -f 3 -d " ");
-	wget -O "$PW/$sample".bam  "$link"; ##this part can be adjusted maybe
+    if [[ ! -f $sample ]]; then
+        pop=$(echo $s_line |cut -f 1 -d " ");
+        link=$(echo $s_line |cut -f 3 -d " ");
+        wget -O "$PW/$sample".bam  "$link"; ##this part can be adjusted maybe
+    fi
 done;
